fix: HSV overflow, app auto-connect, LED live preview
Firmware:
- Fix signed int16_t overflow in hsv_to_rgb (s*remainder for s=255
  exceeds 32767, causing chaotic rainbow colors). Cast to uint16_t.

App:
- Auto-connect on Read/Write/Bootloader (no more 'connect first')
- Live LED preview: loop() polls LED selects and toggle checkboxes,
  sends Feature Report 3 immediately on change
- Sync polling state after read/write to prevent re-send loops

+++ src/board/main.c
@@ -66,9 +66,9 @@ void hsv_to_rgb(uint8_t h, uint8_t s, uint8_t v,
     region    = h / 43;
     remainder = (h - (region * 43)) * 6;
 
-    p = (v * (255 - s)) >> 8;
-    q = (v * (255 - ((s * remainder) >> 8))) >> 8;
-    t = (v * (255 - ((s * (255 - remainder)) >> 8))) >> 8;
+    p = ((uint16_t)v * (255 - s)) >> 8;
+    q = ((uint16_t)v * (255 - (((uint16_t)s * remainder) >> 8))) >> 8;
+    t = ((uint16_t)v * (255 - (((uint16_t)s * (255 - remainder)) >> 8))) >> 8;
 
     switch (region) {
         case 0:  *r = v; *g = t; *b = p; break;
